[X] Test and Debug: SqlLite Dapper / SqlLite EF (fase 1)

+++ KNote_dev/KNote/Model/Result.cs
@@ -14,7 +14,8 @@ namespace KNote.Model
 
         public TEntity Entity { get; set; }
 
-        public int CountColecEntity { get; set; } = 1;
+        //public int CountColecEntity { get; set; } = 1;
+        public long CountColecEntity { get; set; } = 1;
 
         #endregion
